#196 [refact] : 비트맵 변환 함수 util 함수로 확장
